CORS-3479: seting subnet config when not BYO-VPC deployment
Nodes must be correctly distributed among zones in the default
deployment, Manage VPC and no zones is set in the install-config.yaml.

Additionally, when the zones is set, the control plane nodes must follow
the user-defined configuration, placing nodes correctly in the zones,
considering the "subnet internet facing", when the PublicIP is set to
the machine.

This change changes the behavior not letting CAPA decide which zones
will be placed when no subnets are set, keeping the terraform parity
which is distributing control plane nodes between zones, preventing
nodes in the same zone when there are available.