pkg/clusterapi/system: run from metadata
This commit removes the dependency on the installconfig and uses
metadata instead. This will make it easier to restart the capi system
at any point because we will not need to retrieve the installconfig
asset. At the moment, the standalone bootstrap destroy command is the
only example of when we will need to restart the CAPI control plane.