CAPI: Extend bootstrap destroy timeout to 5 min
I have noticed in some cases that the bootstrap destroy timeout
is exceeded at 2 minutes. Let's bump this up to 5 minutes, especially
as there is very little harm in doing so.

This also fixes the logging so that the log message of "Finished
destroying bootstrap resources" is only shown if there is no timeout.