capi provider interface: add bootstrap delete hook
Adds a hook to execute provider-specific behvavior during
bootstrap destroy. This hook should be used to handle any
bootstrap resources that are not deleted in the default
behavior of deleting the capi machine.

The intended first use case of this is to delete aws
security rules for ssh to the bootstrap node. This
use case does not have any relationship to the
machine deletion; so I have opted for the simplest
execution of running this command serially before
the call to destroy the capi machine. If needed,
this command could be run post machine destroy or
even to run concurrently.