AzureMachines: set disk caching to ReadWrite
Change Azure Masters to use ReadWrite cache.

The write is then lazily written to the disk when the cache is flushed
periodically. Customers can additionally force a flush by issuing an
f/sync or fua command.

etcd should be ok with this change since etcd does fsyncs when needed.

https://learn.microsoft.com/en-us/azure/virtual-machines/disks-performance

Based on Terraform changes in:
https://github.com/openshift/installer/pull/8380