OCPBUGS-33745: Move etcd data to its own dir
We save an etcd data dir in order to be able to restart the local
control plane to destroy the bootstrap node. Prior to this commit,
we saved etcd data in the binary directory, but that is cleaned
whenever the control plane shuts down, so it defeats the purpose
of saving the data for a restart.

Instead, save the etcd data in its own dir.