OCPBUGS-33877: bump aws bootstrap destroy timeout
In CI, we are seeing that deleting the SSH rule is frequently hitting
the 5-minute time limit. This increases the time limit to 15 minutes
and adds logging and warning messages for observation.

In testing, most cases of deleting this rule take 10 seconds, but
not infrequently we can see an outlier that takes, say, 8 minutes.