Move CAPI bootstrap destroy hook to after shutdown
Moves the provider hook for bootstrap destroy to run after the CAPI
system has shutdown. AWS is currently the only provider using this
hook. The initial implementation of AWS bootstrap destroy used
CAPI to delete the SSH bootstrap rule, but doing that causes a
full reconciliation of cluster, which is unnecessary when removing
a single rule.

Instead, let's move the bootstrap hook to run after the capi system
has shutdown so we can use the AWS SDK to remove the rule. We need
to do this after shutdown so that there is no chance that CAPI will
try to reconcile the SSH rule as we try to delete it.

In the future, it would be fine to add more hooks to bootstrap destroy
including those that use the CAPI system to delete (in which case we
might want to rename this), but it is not clear that will be needed.