OCPBUGS-33874: capi/aws: delete bootstrap ssh rule with SDK
We do not want to modify the AWSCluster object, specially in the middle
of deleting the bootstrap node, because it triggers a cluster reconcile
loop in CAPA. Instead, let's use the AWS SDK to revoke only the ssh rule
directly.

It should be fine to modify the cluster without CAPA's knowledge because
at the bootstrap destroy stage we're about to shutdown the capi system
anyway. Besides, the Installer is responsible for deleting resources and
it knows how to find the security groups and its rules by the cluster
tag.