OCPBUGS-34295: wait for ip address on machines
Updates the post-machine-creation logic to wait for the IP addresses
to be populated. This ensures that when we write the manifests to disk
the IP address will be present in case we need it for must gather.

Adds debug logging for checking when machines are provisioned
and have obtained the required IP addresses.