OCPBUGS-33661: capi/aws: rename `preserveBootstrapIgnition`
The name doesn't really reflect the purpose of the field, but the
terraform implementation instead. Since the implementation has changed
in capi/capa, let's rename it so users don't expect the ignition object
to not be destroyed when there are enough permissions.

The old field is kept for the deprecation period but will be removed in
the future in favor of the new `BestEffortDeleteIgnition`.