OCPBUGS-34493: baremetal: Don't always enable provisioning-interface.service
In the baremetal bootstrap it is not sufficient to make ironic.service
conditionally Require provisioning-interface.service, as
104eae17b879b84700c31072c3a6eba40fc5e8a1 attempted to do, when the
latter is already enabled because it is WantedBy multi-user.target. As
long as it is enabled, systemd will try to start it.

Do not install provisioning-interface.service, so that it will only be
enabled when ironic.service requires it (i.e. when the provisioning
network is not disabled).

(cherry picked from commit bb91cde94317d1f782b7c511b7532a83028b35fa)