Add empty manifest files to node-joiner config-iso
The script that loads the config ISO only recognises it if all of the
necessary files (defined in GetConfigImageFiles()) are present in the
archive. When creating a node-joiner image, the cluster-deployment and
agent-cluster-install manifests are not actually required (as they are
when creating a new cluster).

To avoid having two different ways of deserializing the config archive,
add empty files for those manifests to keep the script happy. This also
has the effect of creating empty files in the node-joiner ISO, but they
are harmless.