pkg: Allow baremetal platform without MachineAPI capability
There are use cases for using the baremetal platform while disabling MAPI (for example to use CAPI).
The installer will skip rendering BMH, Machine and MachineSet CRs in case MAPI is disabled
This also removes the validation preventing this configuration to allow users
to create the ignition-configs.
The validations still apply for the create cluster target

Fixes METAL-1039

Signed-off-by: Eran Cohen <eranco@redhat.com>