capi system: clean up etcd dir
Etcd data is preserved to support the standalone
openshift-install destroy bootstrap command. We can only delete this
once bootstrap destroy has been completed. Teardown may be called
in other cases, such as an error or user interrupt, so this commit
introduces a separate function to delete the etcd directoy specifically.