pkg/asset/cluster: stop re-entry in capi installs
This commit prevents users from restarting an aborted install,
i.e. from running create cluster multiple times. This check
is a guardrail that parallels the functionality with terraform
installs.

CAPI installs should be re-entrant, and in my experience this
has generally worked fine, but it is not (yet) well tested. Because
this functionality can work and is useful, the commit includes
an environment variable to allow opting in for re-entry.