AWS: Refactor SSH rule bootstrap destroy
The status of the security group rules in the awscluster object
do not seem to be readily updated by CAPA. Instead of inspecting
the awscluster, instead just check directly in AWS as to whether
the rule was deleted.

Furthermore, this checks whether any SSH rule is open to the public
in the control plane security group. Although the installer will
only try to delete the one rule it created. If any additional
SSH rules are present, the installer will identify it and fail.