clusterapi: poll when waiting for provisioning
This commit switches the infrastructure and machine checks to poll
instead of using an exponential backoff. The exponential backoff
is ineffective with our model as the time between checks can become
quite large in later steps. As we are using a local control plane
and there is no concern with rate limiting, a check at a regular
interval is a better solution.