openstack: Attach a bootstrap FIP to the cluster network port
Before this patch, the bootstrap FIP was attached to the first available
port attached to the bootstrap server. As a consequence, the Installer
sometimes tried to attach the FIP to a port that was created on an
additional network.

With this patch, the bootstrap FIP is only attached to the bootstrap
port that exists on the cluster network.