OCPBUGS-35323: aws: delete ignition bucket on bootstrap destroy
Leverage the post-destroy hook to delete the S3 ignition bucket after
the cluster-api system has stopped running.

Because of the `BestEffortDeleteIgnition` option, we need to save its
value to the provider while the cluster-api system is still running
during the bootstrap destroy process. While not ideal, this solution
offers the least impact to consumers of the Installer code as opposed to
saving that value into the metadata.