capi aws: append random string in caller ref
CallerReference needs to be a unique string. We include the infra id,
which is unique, in case that is helpful for human debugging. A random
string of an arbitrary length is appended in case the infra id is reused
which is generally not supposed to happen but does in some edge cases.