OCPBUGS-35494: capi/aws: allow 6443 for private cluster NLB
This is in line with what was done with terraform/aws. The difference
there is that we didn't create a security group for the NLB at all.

With CAPA, we need to allow ingress on 6443 for the NLB for private
cluster installs for cases when, e.g., privateLink is used. Otherwise
the installer won't be able to reach the kube API and the install will
be deemed "failed".