infra/aws: remove apiHost condition on Load Balancer
This is leftover from before CAPA had support for a public LB as the
secondary controlPlane load balancer. We had to configure the AWSCluster
in such a way the primary load balancer would either be `InternetFacing`
if publish was set to "External" or `Internal` otherwise.

Now the primary LB is always `Internal` and the secondary LB only exists
when publish is "External".