images/upi-installer: pin versions of ibmcloud plugins.
The latest version of the vpc-infrastructure plugin has started failing
with
```
$ ibmcloud plugin install vpc-infrastructure -f
Looking up 'vpc-infrastructure' from repository 'IBM Cloud'...
Plug-in 'vpc-infrastructure[infrastructure-service/is] 11.4.0' found in repository 'IBM Cloud'
Attempting to download the binary file...
 24.06 MiB / 24.06 MiB [====================================================] 100.00% 0s
25231360 bytes downloaded
Installing binary...
FAILED
Plug-in directory '/output/.bluemix/plugins/vpc-infrastructure' already exists. Remove the directory and try again.
```
Let's pin to the previous version without this problem. And so that it
doesn't happen with the other plugins, let's pin all of them to specific
versions and just bump when we intend to.