destroy/gcp: set value for `DiscardLocalSsd`
Instance types for which `OnHostMaintenance` is set to `Terminate`, GCP
requires the `DiscardLocalSsd` value to be defined, otherwise we get the
following error when destroying a cluster:
```
WARNING failed to stop instance jiwei-0530b-q9t8w-worker-c-ck6s8 in zone us-central1-c: googleapi: Error 400: VM has a Local SSD attached but an undefined value for `discard-local-ssd`. If using gcloud, please add `--discard-local-ssd=false` or `--discard-local-ssd=true` to your command., badRequest
```

We are setting the value to `true` because we are about to destroy the
cluster, which means destroying the instances and all cluster-owned
resources.