ic/aws: validate BYO instance profiles
Make sure the instance profiles exist and have a role attached. Even
though it's not possible through the AWS console, one can create an
empty instance profile via aws cli.

Because an instance profile without a role is of no use, we make it so
`iamRole` and `iamProfile` parameters are mutually exclusive.