Remove the mitigations parameter in SCOS installations to force kernel defaults
Currently, OKD/SCOS is installed using FCOS boot images. FCOS boot images override the default setting for the mitigations parameter, disabling SMT in Intel-based machines.
This leads the installation to fail with the default machine types in AWS.
This commit will add the KernelArguments stanza to the ignition for the master and worker nodes, in order to remove the kernel parameter mitigations=auto,nosmt.