OCPBUGS-36660 4.17 default channel incorrectly points to stable-4.16
a reoccurring issue, reported by me in the past in OCPBUGS-889, and by @shellyyang1989 in OCPBUGS-26511  
while branching to the next release, default channel must be updated to match, otherwise it will be impossible for the cluster to search for updates in the update graph, unless corrected manually. 

this PR updates the channel of the current master, from which 4.17.0-ec.x builds are created currently.