capi: shutdown system even when capi failed to run
`clusterapi.System().Run()` is not atomic and it can fail after local
controlplane, kube-apiserver, etcd or some controllers are already
running. To make sure the capi system is properly shut down and the etcd
data is cleaned up on errors, we need to `defer` the cleanup before we
even attempt to run the capi system.