CORS-3566: Azure CAPI: disable natgw by setting node subnet id (#8727)
* Azure CAPI: disable natgw by setting node subnet id

Setting the ID on the Subnet disables natgw creation. See:
https://github.com/kubernetes-sigs/cluster-api-provider-azure/blob/21479a9a4c640b43e0bef028487c522c55605d06/api/v1beta1/azurecluster_default.go#L160

Because the node subnet does not already exist, we are using an arbitrary value.
We could populate this with the proper subnet ID in the case of BYO VNET, but
the value currently has no practical effect.

* Set subnet ID to UNKNOWN.

Sets the subnet ID to a clearly unknown value to avoid confusion.