CAPI Azure: gather bootstrap through LB
The Azure bootstrap node cannot have a public IP address, as
Azure does not allow a public IP when the VM is attached to a load
balancer with an outbound rule. Instead, CAPZ creates an inbound
nat rule to allow SSH access through the load balancer.

This commit encapsulates the logic for gathering the bootstrap
host address into its own function and adds a condition to use
the API LB hostname when the bootstrap node cannot have a public ip.

Fixes: OCPBUGS-37540