incorrect folder gen, workaround govmomi vm folder path bug,
This PR resolves an issue when `folder` is undefined
and datacenter exists as a child of a datacenter folder.
instead of using DatacenterFolders.VmFolder we
replace that with a path.Join of the datacenter InventoryPath,
vm and the clusterID. This will provide the entire
correct path to the hidden `/vm/` folder in vCenter.