clusterapi: ignition interface returns secrets
Updates the clusterapi ignition interface so that it returns
all ignition secrets. Prior to this commit, the ignition interface
returned the bootstrap ignition data, and the provision method
turned this data into secrets. Updating the interface to return
all secrets, gives greater flexibility to the platform to completely
control the ignition secrets that are created for that platform.

The motivation is that some platforms such as Nutanix may need
to create per master ignition.