azure: Pick the next available IP address for internal LB
If a cluster is installed in an existing vnet, since CAPI hardcodes
the internal LB IP to 10.0.0.100, there is a chance of conflict
if there was a cluster already installed. Added a check to see
if the default internal LB is available, else pick the next
available IP.

Also added a field in the install config "InternalLBIP" that the
users can set for the internal LB to use.