OCPBUGS-38719: Filter peering zones out
** In GCP, a peering zone can be created and act as a private zone only. Searching for zones based
only on their type (public or private) was causing peering zones to appear. Peering zones are not
currently supported in openshift. When searching for available dns zones filter out the peering zones
if the peering zone config is available (according to the comments, a non null struct indicates a peering zone).