OCPBUGS-36869: Avoid races with provisioning-interface service
When the provisioning network is disabled in Metal IPI, the
provisioning-interface service on the bootstrap can still run to set the
BootstrapProvisioningIP address - however, it does so on the main
(external) network interface. Since this service restarts the interface,
it must not run at the same time as anything that is accessing the
network (e.g. to pull container images) lest those other services be
interrupted. In particular, oneshot services with dependencies must not
fail.

To prevent the network interface being restarted at an inopportune time,
ensure all units have ordering dependencies with
provisioning-interface.service.