Remove copy-pasta from metal3-baremetal-operator service
Some of the configuration in this file has been inherited from the
ironic service. We don't mount ironic.volume, so there is no need to
depend on it. We don't include the /etc/ironic.env environment file, so
we do not need to wait for it to be built and should not attempt to pass
on the HTTP_PORT and IRONIC_KERNEL_PARAMS environment variables from it.

Finally, BMO requires the ironic.service, so set that dependency instead of
using ironic.service's own dependencies. Note that this must be a
Requires dependency because we need the BMO to stop when we stop
ironic.service in master-bmh-update.