baremetal bootstrap: Downgrade Requires dependencies to Wants
Since the image customization service now runs as a kubernetes
controller, it is no longer directly required by ironic (as it was when
it ran on static data and Terraform was doing the provisioning). Ironic
needs only the kernel to be set up by extract-machine-os, which was
previously a transitive dependency.

And although it is wanted by BMO, it should not block BMO startup as BMO
may be able to at least provide some useful debugging information in the
CRDs if the image-customization service is not running.

In turn, the image-customization controller can run and potentially
provide useful error messages if the CoreOS ISO file does not exist, so
use Wants instead of Requires to depend on extract-machine-os.