Remove wait-iptables-init.service
Code to wait for iptables to start was added (to startironic.sh) in the
RHEL 7 era. We kept it because we were not sure it was still needed. We
now know that when oneshot systemd services are restarted, it prevents
any of the services that depend on them from starting. It follows that
the wait-iptables-init service is not having any effect today in RHEL 9,
so it can be removed.