no-jira: bump cluster-api version to 1.8.2
This is a fallout from https://github.com/openshift/installer/pull/9010.

The cluster-api binary version used should be no lower than the
greatest version required by the providers which in this case is 1.8.2
by ibmcloud. Otherwise, we will be running an older binary while linking
the provider against a newer version, possibly requiring new function
args and/or functions.