OCPBUGS-42563: machines: don't sort mpool zones
When sorting the availability zones before creating the CPMS manifest,
we are inadvertently changing the backing array that is used by the MAPI
machine manifests for controlPlane nodes, thus creating a discrepancy
between which zones capi creates the machines in and where they are
assigned to in the mapi manifests.

The CPMS operator sorts the failure domains before using them [1], so
there is no reason for the installer to do that.

[1] https://github.com/openshift/cluster-control-plane-machine-set-operator/blob/611636f48fc87bb09c95121d228697cee72fc118/pkg/machineproviders/providers/openshift/machine/v1beta1/mapping.go#L114-L122