OCPBUGS-30292: If agent proxy defined add RendezvousIP to noProxy
If proxy is defined in install-config, add the RendezvousIP to the
noproxy field in the generated manifests to ensure the Rendezvous
host can be accessed.

This is an additional fix beyond https://github.com/openshift/installer/pull/9083
as it was found that when just setting the IP in 10-default-env.conf it did not
get passed to the assisted-installer when it was started.

(cherry picked from commit 3bff0728696cbc61c0b4abccaf8501963b5e906d)