openstack: Embed CA cert in root credential
If a CA cert is required to talk to your OpenStack then obviously all
services that talk to the cloud need to have both credentials and said
cert. Currently, these users can get their credentials via cloud
credential operator, but they need to source their CA cert from
elsewhere (typically by extracting it from the cloud controller
manager's configuration). This makes configuration of services more
complicated than necessary.

Continue the resolution of the issue by storing the CA cert, if any,
in the root secret on OpenStack. When coupled with the changes
introduced in openshift/cloud-credential-operator#780 [1], this allows
us to dole out the cert to anyone who asks for it via a
'CredentialsRequest'.

[1] https://github.com/openshift/cloud-credential-operator/pull/780

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>