openstack: Migrate generated CCM config to expected format
cluster-cloud-controller-manager-operator has a config syncer that
transforms configuration meant for the legacy OpenStack cloud-provider
to the former expected by the modern, out-of-tree cloud-provider [1]. We
are currently still generating configuration for the former, which means
this config syncer continues to need to transform the configuration file
format. Resolve this by updating the generated configuration to match
what we actually expect nowadays.

[1] https://github.com/openshift/cluster-cloud-controller-manager-operator/blob/release-4.17/pkg/cloud/openstack/openstack.go#L156-L190

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>