Create public API for CheckIPAddressAvailability
Prior to d02d7ee7b5e06b933f50e7aa23e8412d8b33cfc8, the
getVirtualNetworksClient() function was private to the client, and only
used by public functions that are part of the API interface. That change
made the function public without making it part of the API interface.

Instead, add a new function to the API interface that does what we want
to do, and return getVirtualNetworksClient() to being an internal
implementation detail.