[release-4.15] ARO-12457: data/data: include bootkube files in go module (#9243)
* data/data: include bootkube files in go module

Adds a data/data package and uses a side effect of //go:embed to
include specific data directories in our go module, so that the ARO
wrapper can vendor these static assets. This is intended to be
temporary. The installer does not use this package. Once ARO has
moved off the wrapper (to use the openshift-install binary directly)
this file and package can be removed.

* fixup! data/data: include bootkube files in go module

---------

Co-authored-by: Patrick Dillon <padillon@redhat.com>