Remove narrow timeout from etcd bootstrap member removal gate.
Waiting for the etcd bootstrap member to be removed from the etcd cluster has a timeout of five
minutes, unless the overall timeout is reached first. This was apparently sufficient for most cases
when the check was only in effect for single-node clusters, but it occasionally times out on HA
clusters. Enforcing a timeout for this step alone is fragile. It is important to guarantee that the
bootstrap resources are not torn before the etcd bootstrap member has been removed from the etcd
cluster. The time spent waiting for it to happen will fluctuate based on how long it takes for the
etcd operator to observe that it is safe to proceed without losing quorum.