OCPBUGS-45669: aws: fix Node Port Service rule removal
This is a fallout from https://github.com/openshift/installer/pull/9281

By including the Ipv6 params I've made it so AWS would only return a
result if a rule satisfying both the Ipv4 and Ipv6 params was present,
which is not the case currently.

Dropping the Ipv6 part as Ipv6 is not supported in 4.16.