OCPBUGS-47477: aws: add supported zones to skip undesired
Add the method to skip undesired zones when discoverying zones from
metadata.

This change targets explicity to skup the zone us-east-1e which provides
very limited offerings of EC2, with no EC2 supported by installer.

This change prevents creating infrastructure in such zone as it will
not be supported to run workloads (create instances) on it.