OCPBUGS-46596: aws/perm - default ec2:DescribeInstanceTypeOfferings
ec2:DescribeInstanceTypeOfferings is used by machine pools to discover
supported instance type in the region and zones, when it isn't set in
the pool (control plane, compute, or edge).

The discover falls back to the m6i which is supported in mostly regions,
althrough some regions (e.g. ap-southeast-4 and eu-south-2) will fail as
that tye isn't supported. To the discover mechanism works properly
globally the permission must be added by default.

This permission is missing since minimum permissions has been
introduced, currently only the edge pool is including this permission.
This change moves that requirement to default create group.