CORS-3295: hack: add script to verify capi CRDs (#8106)
* CORS-3295: hack: add script to verify capi CRDs

This script will be used to create a test in CI to check that the capi
CRDs in data/data/cluster-api match what's expected from the provider
version used in cluster-api/providers.

As a side effect, this script can be used to generate the CRD, allowing
a dev to then commit the result.

* hack: add capi CRD generation to verify-capi-manifests script

* hack/verify-capi-manifests: use Golang 1.22 image

* hack: add azureaso to the manifest generation

Although we are not technically generating the manifest, only
downloading it from the upstream repo.

* data/data/cluster-api: reconcile CRDs differences

This is the diff generated by running `hack/verify-capi-manifests.sh`