azure: Add check for asset before destroy
Adding a check to see if PlatformProvisionCheck asset exists
before initiating destroy cluster. This asset is only created
during create cluster phase which means the user used that command
to create the cluster and hence can be cleaned with no
consequences.

In the case of UPI, deleting a resource group would possibly
delete other resources in the rg which might've been created
by the user and could be something we shouldn't delete.