VSphere: add vips to machinenetwork
Appends the ingress and api vips to the machinenetworks in the cluster
infrastructure object, because the cluster network operator checks to
ensure that the machinenetwork contains those vips. In some edge
cases we have seen valid configurations fail when the VIPs appear
to be outside the valid range.

Uses k8s.io/utils/net package to determine if a string
ip address is ipv6.