Azure cluster manifest: add azure stack fork
CAPZ expects the azure environment name to be HybridEnvironment,
rather than the azurestackcloud value we use. So this commit
translates the environment when populating the cluster manifest.

This commit uses the capz azurestack fork when generating the cluster
manifest. The only API difference between the fork and CAPZ is a single
field for ARMEndpoint. So this dumbly converts the capz.AzureCluster
into the forked equivalent and then sets the ARMEndpoint field.